Change the C types for stdint types to prevent damaged on non-32-Bits architectures
git-svn-id: https://svn.code.sf.net/p/splix/code/splix@20 10358e0b-e31b-0410-b13d-bf5c54612e97

+++ include/band.h
@@ -21,6 +21,8 @@
 #ifndef BAND_H_
 #define BAND_H_
 
+#include <sys/types.h>
+#include <stdint.h>
 
 class Band
 {
@@ -32,8 +34,8 @@ class Band
 		unsigned long	_clipping;
 		
 	protected:
-		unsigned char*	_algorithm0(unsigned long *size);
-		unsigned char*	_algorithm11(unsigned long *size);
+		unsigned char*	_algorithm0(size_t *size);
+		unsigned char*	_algorithm11(size_t *size);
 
 	public:
 		Band(unsigned long bandWidth, unsigned long bandHeight);
@@ -54,7 +56,7 @@ class Band
 		void		clean() {_line = 0;}
 		int		addLine(unsigned char *line, unsigned long width);
 
-		unsigned char*	exportBand(int algorithm, unsigned long *size);
+		unsigned char*	exportBand(int algorithm, size_t *size);
 
 		bool		isFull() const 
 				{return _line == _height ? true : false;}
